Adding a basic first pass at getting a core going.
I've added discoverers (these find projects), projects (these define the
functions for each type of project), and plugins (these define extra
functionality that can be toggled on inside of a project).

I have also added functionality to discover the root of the project tree
(uses a .shantyroot file, will be created when `shanty init` is run) and
executing configuration of a project using a Shantyfile as found inside
the root of a project.