Use TaskEnv to pass around data to tasks, add config.require support.
We want the graph to be lazily loaded, so I've added TaskEnv, a monad
style class that encapsulates the loading of the config, graph and other
things. This is the only object we pass to tasks other than the options,
so they can use this to get the graph and config.

In addition, the first config option called "require" has been added.
This is a config option that takes a list of files, folders or
shell-globs that should be required when Shanty is loaded. This is used
to load custom build-scripts or project types from within a repo that
aren't proper packaged plugins or the like.

Tasks have been renamed to TaskSets to better distinguish the difference
between the classes that have tasks in them, and the tasks themselves.

Finally, I have upgraded the Ruby version on Travis to 2.2.