Allow reconciling of two ProjectTemplates for the same directory.
This problem originally manifested because we want to have a discover
for Shantyfiles (in case those projects aren't discovered by anything
else), but we want to make sure the Shantyfile discoverer doesn't win
over a more specific discovered template.

A priority attribute on ProjectTemplate was introduced, and we now sort
by this (larger is better) and uniq (ie. largest priority or last one in
wins).