Remove the notion of a project type.
Rather than a "type" of a project, plugins start to emerge as the "one
true way™" of doing everything. We will have a single project class, and
one would mix in plugins to add the needed functionality. There are some
new methods on the graph to find projects with a given plugin, replacing
the old `all_by_type` and `changed_by_type`: these new methods are
`all_with_plugin` and `changed_with_plugin`, respectively.

This work is in regards to #5.#5.#5.